Deploys site with GitHub Pages
Migrates the deployment process to use GitHub Pages.

This change updates the workflow to build the site using mkdocs and deploy it directly to GitHub Pages, leveraging the `actions/deploy-pages` action. It also adds steps to compress and embed fonts into the PDF and upload the PDF as an artifact.